feat(userRegistration): write descriptive signup validation messages
- write messgaes for non-alphanumeric characters
- descriptive message if the user exists
- ensure password is 8 or more characters
- ensure that the email entered is valid

[Finishes #162949071]
