+++ authors/apps/authentication/urls.py
@@ -2,7 +2,6 @@ from django.urls import path
 from .views import (
     LoginAPIView, RegistrationAPIView, UserRetrieveUpdateAPIView
 )
-
 app_name = 'authentication'
 urlpatterns = [
     path('user/', UserRetrieveUpdateAPIView.as_view()),
